+++ packages/embedded-client/src/logger.ts
@@ -3,25 +3,21 @@ import { monotonicFactory } from "ulid";
 
 import type { AuditTrail } from "@cerbos/api/cerbos/audit/v1/audit_pb";
 import type {
-  CheckResourcesResponse as CheckResourcesResponseWithAuditTrail,
-  PlanResourcesResponse as PlanResourcesResponseWithAuditTrail,
+  CheckResourcesRequestValid as ServerCheckResourcesRequest,
+  CheckResourcesResponse as ServerCheckResourcesResponse,
+  PlanResourcesRequestValid as ServerPlanResourcesRequest,
+  PlanResourcesResponse as ServerPlanResourcesResponse,
 } from "@cerbos/api/cerbos/cloud/epdp/v2/epdp_pb";
-import type { CheckOutput_ActionEffect } from "@cerbos/api/cerbos/engine/v1/engine_pb";
 import type {
-  CheckResourcesRequestValid,
-  PlanResourcesRequestValid,
+  CheckResourcesRequestValid as ClientCheckResourcesRequest,
+  PlanResourcesRequestValid as ClientPlanResourcesRequest,
 } from "@cerbos/api/cerbos/request/v1/request_pb";
 import type {
-  CheckResourcesResponse,
-  PlanResourcesResponse,
+  CheckResourcesResponseValid as ClientCheckResourcesResponse,
+  PlanResourcesResponseValid as ClientPlanResourcesResponse,
 } from "@cerbos/api/cerbos/response/v1/response_pb";
 import type {
-  CheckInput,
-  CheckOutput,
   DecisionLogEntryMethod,
-  DecodedAuxData,
-  PlanResourcesInput,
-  PlanResourcesOutput,
   PolicySource,
   RequestContext,
 } from "@cerbos/core";
@@ -32,7 +28,6 @@ import {
   planResourcesInputFromProtobuf,
   planResourcesOutputFromProtobuf,
   requestContextFromProtobuf,
-  requireField,
 } from "@cerbos/core/~internal";
 
 import type { Options } from "./options.js";
@@ -49,54 +44,58 @@ export class DecisionLogger {
   }
 
   public async logCheckResources(
-    request: CheckResourcesRequestValid,
-    auxData: DecodedAuxData | undefined,
     headers: Headers,
     bundleId: string,
-    response: CheckResourcesResponseWithAuditTrail | undefined,
+    clientRequest: ClientCheckResourcesRequest,
+    serverRequest: ServerCheckResourcesRequest,
+    serverResponse: ServerCheckResourcesResponse | undefined,
+    clientResponse: ClientCheckResourcesResponse | undefined,
     error: unknown,
   ): Promise<void> {
     const callId = await this.logDecision(
       {
         name: "CheckResources",
-        inputs: checkInputs(request, auxData),
-        outputs: checkOutputs(response?.response),
+        inputs: serverRequest.inputs.map(checkInputFromProtobuf),
+        outputs: serverResponse?.outputs.map(checkOutputFromProtobuf) ?? [],
         error: errorMessage(error),
       },
-      requestContext(request),
-      response?.auditTrail,
+      requestContext(clientRequest),
+      serverResponse?.auditTrail,
       headers,
       bundleId,
     );
 
-    if (response?.response) {
-      response.response.cerbosCallId = callId;
+    if (clientResponse) {
+      clientResponse.cerbosCallId = callId;
     }
   }
 
   public async logPlanResources(
-    request: PlanResourcesRequestValid,
-    auxData: DecodedAuxData | undefined,
     headers: Headers,
     bundleId: string,
-    response: PlanResourcesResponseWithAuditTrail | undefined,
+    clientRequest: ClientPlanResourcesRequest,
+    serverRequest: ServerPlanResourcesRequest,
+    serverResponse: ServerPlanResourcesResponse | undefined,
+    clientResponse: ClientPlanResourcesResponse | undefined,
     error: unknown,
   ): Promise<void> {
     const callId = await this.logDecision(
       {
         name: "PlanResources",
-        input: planResourcesInput(request, auxData),
-        output: planResourcesOutput(request, response?.response),
+        input: planResourcesInputFromProtobuf(serverRequest.input),
+        output:
+          serverResponse?.output &&
+          planResourcesOutputFromProtobuf(serverResponse.output),
         error: errorMessage(error),
       },
-      requestContext(request),
-      response?.auditTrail,
+      requestContext(clientRequest),
+      serverResponse?.auditTrail,
       headers,
       bundleId,
     );
 
-    if (response?.response) {
-      response.response.cerbosCallId = callId;
+    if (clientResponse) {
+      clientResponse.cerbosCallId = callId;
     }
   }
 
@@ -130,105 +129,6 @@ export class DecisionLogger {
   }
 }
 
-function checkInputs(
-  request: CheckResourcesRequestValid,
-  auxData: DecodedAuxData | undefined,
-): CheckInput[] {
-  return request.resources.map<CheckInput>(({ resource, actions }) => ({
-    ...checkInputFromProtobuf({
-      $typeName: "cerbos.engine.v1.CheckInput",
-      requestId: request.requestId,
-      principal: request.principal,
-      resource,
-      actions,
-    }),
-    auxData,
-  }));
-}
-
-function checkOutputs(
-  response: CheckResourcesResponse | undefined,
-): CheckOutput[] {
-  if (!response) {
-    return [];
-  }
-
-  return response.results.map<CheckOutput>((result) => {
-    requireField(
-      "CheckResourcesResponse.ResultEntry.resource",
-      result.resource,
-    );
-
-    requireField("CheckResourcesResponse.ResultEntry.meta", result.meta);
-
-    const actions: Record<string, CheckOutput_ActionEffect> = {};
-
-    for (const [action, effect] of Object.entries(result.actions)) {
-      const meta = result.meta.actions[action];
-
-      actions[action] = {
-        $typeName: "cerbos.engine.v1.CheckOutput.ActionEffect",
-        effect,
-        policy: meta?.matchedPolicy ?? "",
-        scope: meta?.matchedScope ?? "",
-      };
-    }
-
-    return checkOutputFromProtobuf({
-      $typeName: "cerbos.engine.v1.CheckOutput",
-      requestId: response.requestId,
-      resourceId: result.resource.id,
-      actions,
-      outputs: result.outputs,
-      effectiveDerivedRoles: result.meta.effectiveDerivedRoles,
-      validationErrors: result.validationErrors,
-      evaluationErrors: []
-    });
-  });
-}
-
-function planResourcesInput(
-  request: PlanResourcesRequestValid,
-  auxData: DecodedAuxData | undefined,
-): PlanResourcesInput {
-  return {
-    ...planResourcesInputFromProtobuf({
-      $typeName: "cerbos.engine.v1.PlanResourcesInput",
-      requestId: request.requestId,
-      principal: request.principal,
-      resource: request.resource,
-      action: request.action, // eslint-disable-line @typescript-eslint/no-deprecated
-      actions: request.actions,
-      includeMeta: request.includeMeta,
-    }),
-    auxData,
-  };
-}
-
-function planResourcesOutput(
-  request: PlanResourcesRequestValid,
-  response: PlanResourcesResponse | undefined,
-): PlanResourcesOutput | undefined {
-  if (!response) {
-    return undefined;
-  }
-
-  return planResourcesOutputFromProtobuf({
-    $typeName: "cerbos.engine.v1.PlanResourcesOutput",
-    requestId: response.requestId,
-    filter: response.filter,
-    filterDebug: response.meta?.filterDebug ?? "",
-    action: response.action, // eslint-disable-line @typescript-eslint/no-deprecated
-    actions: response.actions,
-    policyVersion: response.policyVersion,
-    scope: request.resource.scope,
-    evaluationErrors: [],
-    validationErrors: response.validationErrors,
-    kind: response.resourceKind,
-    matchedScopes: response.meta?.matchedScopes ?? {},
-  });
-}
-
 function errorMessage(error: unknown): string | undefined {
   if (error === undefined) {
     return undefined;
@@ -255,7 +155,7 @@ function metadata(headers: Headers): Record<string, string[]> {
 
 function requestContext({
   requestContext,
-}: CheckResourcesRequestValid | PlanResourcesRequestValid):
+}: ClientCheckResourcesRequest | ClientPlanResourcesRequest):
   RequestContext | undefined {
   return requestContext && requestContextFromProtobuf(requestContext);
 }